GitHub migration
This is the first playable version of the project. It has already been fully developed offline within a local Unity setup, and this update is the migration of said version.
It contains a chess board with a full set of playable pieces following the standard rules of chess. From now on, every additional functionality will be added directly here, including additional testing branches.