Added online functionality and first build
Two new buttons have been added, one for creating a host instance and one for joining as a client. Once a host and a client have been determined, the board sets up and the game logic begins working.

Also, a developer build folder has been added to the repository to ease the tests on network features and future problem solving.