Fixed client unable to move
Added RPC calls to OnMouseUp() of Move Plates to ask the Host permission to make a move. Now, the client can move freely as well as the host and both moves are reflected for both players in real time.

New bug has appeared when calculating move plates. Needs to be addressed.