New mechanic: Promoting
Now pawns will promote when reaching the last rank. Pawns usually can promote to any piece other than a pawn and a king, but here it will always automatically promote to a queen. This is done mainly for game pacing issues, since the game has no turns, stopping the game to promote would be a hustle to the player and you almost never want to promote to any other piece. This could change in a future update depending on the direction the game takes.