New mechanic: Player division
More so than a mechanic this is a fix. Previously you could control your opponents pieces as well as yours. Although this could've made for an interesting mechanic, it's not the intended use of the pieces. Now, white pieces will only be moveable by the host and black pieces will only be moveable by the client.
Also, when clicking on a piece of the opposite team, a debug.log will pop notifying this and naming the correct team of the player that clicked to ensure the roles are consistent.