New mechanic: Game over detection and New Game functionality
The first closed gameplay loop has been added. Now, when a king is taken the game ends declaring a winner. A new panel pops up announcing the winner and it contains a button labeled "New Game" that restarts the board when pressed and a new game takes start.

Also, a bug has been fixed where the pieces would appear on top of the UI. Now they have their own sorting layer and appear in the background with the board.