Added main menus (+ game cleanup logic)
Now when entering the game a main menu is presented with the logo of the game, an exit button and a play button that brings up another menu. There you can select the gamemode of choice or go back. Gamemodes are classical chess (yet to be implemented) and the main variant, now named "Chess Kune Do". When clicking on the latter, the game will bring up the usual screen with an added button at the top left that lets you return to the main menu, cleaning the board and the network connections.

Introduced MenuController.cs to manage menu navigation and game session transitions. Moved network connection logic from Game.cs to a new StartGameSetup method and added CleanupGame to reset game state when returning to the menu.